telegram_scraper/requirements.txt
# Core dependencies for Telegram scraping
telethon>=1.28.0
python-dotenv>=1.0.0

# For perceptual hashing (image deduplication)
imagehash>=4.3.0
Pillow>=9.0.0

# For async HTTP (geocoding, etc.)
httpx>=0.24.0

# For language detection
langdetect>=1.0.9

# Optional: For LLM entity extraction
# anthropic>=0.18.0

# Data handling
pandas>=1.5.0
